+++ packages/at_commons/lib/src/verb/enroll_params.g.dart
@@ -1,4 +1,5 @@
 // GENERATED CODE - DO NOT MODIFY BY HAND
+// dart run build_runner build to generate this file
 
 part of 'enroll_params.dart';